move model classes to submodule
