+++ custom_components/sma_ennexos/sma/model/errors.py
@@ -1,0 +1,17 @@
+"""SMA API error classes."""
+
+
+class SMAApiClientError(Exception):
+    """Exception to indicate a general API error."""
+
+
+class SMAApiCommunicationError(SMAApiClientError):
+    """Exception to indicate a communication error."""
+
+
+class SMAApiAuthenticationError(SMAApiClientError):
+    """Exception to indicate an authentication error."""
+
+
+class SMAApiParsingError(SMAApiClientError):
+    """Exception to indicate a parsing error."""
